Struct rand_isaac::isaac64::Isaac64Rng

pub struct Isaac64Rng(_);

A random number generator that uses ISAAC-64, the 64-bit variant of the ISAAC algorithm.

ISAAC stands for "Indirection, Shift, Accumulate, Add, and Count" which are the principal bitwise operations employed. It is the most advanced of a series of array based random number generator designed by Robert Jenkins in 19961.

ISAAC-64 is mostly similar to ISAAC. Because it operates on 64-bit integers instead of 32-bit, it uses twice as much memory to hold its state and results. Also it uses different constants for shifts and indirect indexing, optimized to give good results for 64bit arithmetic.

ISAAC-64 is notably fast and produces excellent quality random numbers for non-cryptographic applications.

In spite of being designed with cryptographic security in mind, ISAAC hasn't been stringently cryptanalyzed and thus cryptographers do not not consensually trust it to be secure. When looking for a secure RNG, prefer Hc128Rng instead, which, like ISAAC, is an array-based RNG and one of the stream-ciphers selected the by eSTREAM contest.

Overview of the ISAAC-64 algorithm:

(in pseudo-code)

Input: a, b, c, s[256] // state
Output: r[256] // results

mix(a,i) = !(a ^ a << 21)  if i = 0 mod 4
             a ^ a >>  5   if i = 1 mod 4
             a ^ a << 12   if i = 2 mod 4
             a ^ a >> 33   if i = 3 mod 4

c = c + 1
b = b + c

for i in 0..256 {
    x = s_[i]
    a = mix(a,i) + s[i+128 mod 256]
    y = a + b + s[x>>3 mod 256]
    s[i] = y
    b = x + s[y>>11 mod 256]
    r[i] = b
}

This implementation uses BlockRng64 to implement the RngCore methods.

See for more information the documentation of IsaacRng.

Bob Jenkins, ISAAC and RC4 ↩

Methods

impl Isaac64Rng

pub fn new_from_u64(seed: u64) -> Self

Deprecated since 0.6.0
: use SeedableRng::seed_from_u64 instead

Create an ISAAC-64 random number generator using an u64 as seed. If seed == 0 this will produce the same stream of random numbers as the reference implementation when used unseeded.